- MFH  - add 64bit typedef support  - fix vc6 build (1/2)  - still broken/does not build cleanly, more to come
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@268430 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/fileinfo/libmagic/softmagic.c
@@ -582,16 +582,16 @@ cvt_64(union VALUETYPE *p, const struct magic *m)
 	if (m->num_mask) \
 		switch (m->mask_op & FILE_OPS_MASK) { \
 		case FILE_OPADD: \
-			p->fld += cast m->num_mask; \
+			p->fld += cast (int64_t)m->num_mask; \
 			break; \
 		case FILE_OPMINUS: \
-			p->fld -= cast m->num_mask; \
+			p->fld -= cast (int64_t)m->num_mask; \
 			break; \
 		case FILE_OPMULTIPLY: \
-			p->fld *= cast m->num_mask; \
+			p->fld *= cast (int64_t)m->num_mask; \
 			break; \
 		case FILE_OPDIVIDE: \
-			p->fld /= cast m->num_mask; \
+			p->fld /= cast (int64_t)m->num_mask; \
 			break; \
 		} \
 
@@ -1865,7 +1865,7 @@ magiccheck(struct magic_set *ms, struct magic *m)
 	case 'x':
 		if ((ms->flags & MAGIC_DEBUG) != 0)
 			(void) fprintf(stderr, "%llu == *any* = 1\n",
-			    (unsigned long long)v);
+			    (uint64_t)v);
 		matched = 1;
 		break;
 
@@ -1873,7 +1873,7 @@ magiccheck(struct magic_set *ms, struct magic *m)
 		matched = v != l;
 		if ((ms->flags & MAGIC_DEBUG) != 0)
 			(void) fprintf(stderr, "%llu != %llu = %d\n",
-			    (unsigned long long)v, (unsigned long long)l,
+			    (uint64_t)v, (uint64_t)l,
 			    matched);
 		break;
 
@@ -1881,7 +1881,7 @@ magiccheck(struct magic_set *ms, struct magic *m)
 		matched = v == l;
 		if ((ms->flags & MAGIC_DEBUG) != 0)
 			(void) fprintf(stderr, "%llu == %llu = %d\n",
-			    (unsigned long long)v, (unsigned long long)l,
+			    (uint64_t)v, (uint64_t)l,
 			    matched);
 		break;
 
@@ -1890,14 +1890,14 @@ magiccheck(struct magic_set *ms, struct magic *m)
 			matched = v > l;
 			if ((ms->flags & MAGIC_DEBUG) != 0)
 				(void) fprintf(stderr, "%llu > %llu = %d\n",
-				    (unsigned long long)v,
-				    (unsigned long long)l, matched);
+				    (uint64_t)v,
+				    (uint64_t)l, matched);
 		}
 		else {
 			matched = (int64_t) v > (int64_t) l;
 			if ((ms->flags & MAGIC_DEBUG) != 0)
 				(void) fprintf(stderr, "%lld > %lld = %d\n",
-				    (long long)v, (long long)l, matched);
+				    (uint64_t)v, (uint64_t)l, matched);
 		}
 		break;
 
@@ -1906,14 +1906,14 @@ magiccheck(struct magic_set *ms, struct magic *m)
 			matched = v < l;
 			if ((ms->flags & MAGIC_DEBUG) != 0)
 				(void) fprintf(stderr, "%llu < %llu = %d\n",
-				    (unsigned long long)v,
-				    (unsigned long long)l, matched);
+				    (uint64_t)v,
+				    (uint64_t)l, matched);
 		}
 		else {
 			matched = (int64_t) v < (int64_t) l;
 			if ((ms->flags & MAGIC_DEBUG) != 0)
 				(void) fprintf(stderr, "%lld < %lld = %d\n",
-				       (long long)v, (long long)l, matched);
+				       (int64_t)v, (int64_t)l, matched);
 		}
 		break;
 
@@ -1921,16 +1921,16 @@ magiccheck(struct magic_set *ms, struct magic *m)
 		matched = (v & l) == l;
 		if ((ms->flags & MAGIC_DEBUG) != 0)
 			(void) fprintf(stderr, "((%llx & %llx) == %llx) = %d\n",
-			    (unsigned long long)v, (unsigned long long)l,
-			    (unsigned long long)l, matched);
+			    (uint64_t)v, (uint64_t)l,
+			    (uint64_t)l, matched);
 		break;
 
 	case '^':
 		matched = (v & l) != l;
 		if ((ms->flags & MAGIC_DEBUG) != 0)
 			(void) fprintf(stderr, "((%llx & %llx) != %llx) = %d\n",
-			    (unsigned long long)v, (unsigned long long)l,
-			    (unsigned long long)l, matched);
+			    (uint64_t)v, (uint64_t)l,
+			    (uint64_t)l, matched);
 		break;
 
 	default:
